refactor!: drop maxRedirect
