+++ lib/api/api-pipeline.js
@@ -13,7 +13,6 @@ const {
   RequestAbortedError
 } = require('../core/errors')
 const util = require('../core/util')
-const RedirectHandler = require('../handler/redirect-handler')
 const { addSignal, removeSignal } = require('./abort-signal')
 
 const kResume = Symbol('resume')
@@ -241,15 +240,7 @@ class PipelineHandler extends AsyncResource {
 function pipeline (opts, handler) {
   try {
     const pipelineHandler = new PipelineHandler(opts, handler)
-    if (opts?.maxRedirections != null && (!Number.isInteger(opts?.maxRedirections) || opts?.maxRedirections < 0)) {
-      throw new InvalidArgumentError('maxRedirections must be a positive number')
-    }
-
-    if (opts?.maxRedirections > 0) {
-      RedirectHandler.buildDispatch(this, opts.maxRedirections)({ ...opts, body: pipelineHandler.req }, pipelineHandler)
-    } else {
-      this.dispatch({ ...opts, body: pipelineHandler.req }, pipelineHandler)
-    }
+    this.dispatch({ ...opts, body: pipelineHandler.req }, pipelineHandler)
     return pipelineHandler.ret
   } catch (err) {
     return new PassThrough().destroy(err)
