+++ lib/api/api-stream.js
@@ -5,7 +5,6 @@ const { finished, PassThrough } = require('node:stream')
 const { AsyncResource } = require('node:async_hooks')
 const { InvalidArgumentError, InvalidReturnValueError } = require('../core/errors')
 const util = require('../core/util')
-const RedirectHandler = require('../handler/redirect-handler')
 const { getResolveErrorBodyCallback } = require('./util')
 const { addSignal, removeSignal } = require('./abort-signal')
 
@@ -209,15 +208,6 @@ function stream (opts, factory, callback) {
 
   try {
     const handler = new StreamHandler(opts, factory, callback)
-
-    if (opts?.maxRedirections != null && (!Number.isInteger(opts?.maxRedirections) || opts?.maxRedirections < 0)) {
-      throw new InvalidArgumentError('maxRedirections must be a positive number')
-    }
-
-    if (opts?.maxRedirections > 0) {
-      RedirectHandler.buildDispatch(this, opts.maxRedirections)(opts, handler)
-      return
-    }
 
     this.dispatch(opts, handler)
   } catch (err) {
