+++ lib/interceptor/redirect.js
@@ -7,12 +7,12 @@ function createRedirectInterceptor ({ maxRedirections: defaultMaxRedirections } 
     return function Intercept (opts, handler) {
       const { maxRedirections = defaultMaxRedirections, ...rest } = opts
 
-      if (maxRedirections == null) {
+      if (maxRedirections == null || maxRedirections === 0) {
         return dispatch(opts, handler)
       }
 
-      const redirectHandler = new RedirectHandler(dispatch, maxRedirections, opts, handler)
       const dispatchOpts = { ...rest, maxRedirections: 0 } // Stop sub dispatcher from also redirecting.
+      const redirectHandler = new RedirectHandler(dispatch, maxRedirections, dispatchOpts, handler)
       return dispatch(dispatchOpts, redirectHandler)
     }
   }
